requirements.txt
# 🦁 ATHLYNX AI - Python Dependencies
# Athlynx-AI-Start-Up-Launch-All-Phase-Beginning-Phase-1-2026-#14
# 100% Python + Julia Stack

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Templates
jinja2==3.1.3
aiofiles==23.2.1

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0

# AWS Services
boto3==1.34.25
botocore==1.34.25

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.0

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Data Processing
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Utilities
email-validator==2.1.0
python-dateutil==2.8.2

# Logging & Monitoring
structlog==24.1.0
sentry-sdk==1.39.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Production
gunicorn==21.2.0

# Dreams Do Come True 2026! 🚀
